requirements.txt
# requirements.txt

certifi==2023.11.17
charset-normalizer==3.3.2
colorlog==6.7.0
deepl==1.16.1
idna==3.6
packaging==23.2
Pillow==10.1.0
PyMonCtl==0.7
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
pytesseract==0.3.10
python-xlib==0.33
PyWinBox==0.6
PyWinCtl==0.3
requests==2.31.0
shiboken6==6.6.0
six==1.16.0
typing_extensions
urllib3==2.1.0
